requirements.txt
flask
pdfplumber
sentence-transformers
scikit-learn
werkzeug
google-genai
python-dotenv
reportlab
textblob
nltk